Pull out extern.ts
