+++ contracts/assemblyscript-poc/contract/src/cosmwasm-std/extern.ts
@@ -1,0 +1,22 @@
+import { toUtf8 } from "../cosmwasm-encoding";
+import { allocate, deallocate, keepOwnership, readRegion } from "./cosmwasm";
+import * as env from "./env";
+
+export class Extern {
+  // eslint-disable-next-line no-shadow
+  constructor(public readonly canonicalize: (humanAddress: string) => Uint8Array) {}
+}
+
+export function canonicalize(human: string): Uint8Array {
+  const humanEncoded = toUtf8(human);
+  const resultPtr = allocate(50);
+  const returnCode = env.canonicalize_address(keepOwnership(humanEncoded), resultPtr);
+  if (returnCode < 0) {
+    throw new Error(
+      "Call to env.canonicalize_address failed with return code " + returnCode.toString(),
+    );
+  }
+  const canonical = readRegion(resultPtr);
+  deallocate(resultPtr);
+  return canonical;
+}
