+++ contracts/assemblyscript-poc/contract/src/cosmwasm-std/index.ts
@@ -1,1 +1,2 @@
 export * from "./cosmwasm";
+export { Extern, canonicalize } from "./extern";
